refactor(scripts): Share AWS login check
Since it's now used in more than 1 place.

TOWARDS PLA-209
